A test signal of the phase coherence between three oscillators

Generate test signals which involve three oscillators described in Kim and Powers (1979).

kim_and_powers_model(
  fbfN = 0.22,
  fcfN = 0.375,
  fdfN = fbfN + fcfN,
  num_points = 128,
  num_records = 64,
  noise_sd = 0.1,
  phase_coherence = TRUE,
  product_term = FALSE
)

Arguments

fbfN

b's frequency divided by the Nyquist frequency; 0.220 by default.

fcfN

c's frequency divided by the Nyquist frequency; 0.375 by default.

fdfN

d's frequency divided by the Nyquist frequency; fbfN + fcfN by default.

num_points

The number of sampling points in a record; 128 by default.

num_records

The number of records; 64 by default.

noise_sd

The standard deviation of a Gaussian noise perturbing samples; 0.1 (-20dB) by default.

phase_coherence

If TRUE (default), the phase coherence in the signal d is on; otherwise off.

product_term

If TRUE, the product of b and c is included in the model; FALSE by default.

Value

A matrix of num_points rows x num_records columns.

Details

This function produces a list of numeric vectors; its each element represents a test signal in which three oscillators b, c, and d are superimposed. The ratio of the frequency of b (f1) to the Nyquist frequency is 0.220 and the ratio of the frequency of c (f2) to the Nyquist frequency is 0.375, by default. The d's frequency f3 is equal to f1 + f2 unless specified otherwise. Optionally the product of b and c is also added to signals.

Examples

data <- kim_and_powers_model()